Move compression levels to zim.h
